% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_manipulations.R
\name{INLA2long}
\alias{INLA2long}
\title{Convert list of parameter marginals to data frame}
\usage{
INLA2long(inla_res, type = "fixed")
}
\arguments{
\item{inla_res}{}

\item{type}{"fixed" or "hyperpar"}
}
\value{

}
\description{
Convert list of parameter marginals to data frame
}
