% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotMarginals}
\alias{PlotMarginals}
\title{Plot posterior marginals from INLA model}
\usage{
PlotMarginals(margs_df)
}
\arguments{
\item{margs_df}{}
}
\value{

}
\description{
Plot posterior marginals from INLA model
}
