% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_manipulations.R
\name{ReturnMarginalDF}
\alias{ReturnMarginalDF}
\title{Convert all parameter marginals for INLA object to data frame}
\usage{
ReturnMarginalDF(inla_res)
}
\arguments{
\item{inla_res}{}
}
\value{

}
\description{
Convert all parameter marginals for INLA object to data frame
}
